requirements.txt
pdbpp 
backtrader 
python.dotenv 
numpy 
pandas 
mysql-connector-python 
sqlalchemy